Attempt to load any available RMW implementation. (#189)
* Attempt to load any available RMW implementation.

The old logic was to load what the user asked for (via the
RMW_IMPLEMENTATION environment variable).  If the user didn't
explicitly ask for an implementation, then it would attempt
to load the default RMW implementation.

The new logic is to load what the user asked for.  If the user
didn't explicitly ask for an implementation, attempt to load
the default RMW.  If that fails for some reason, fall back to
loading any available RMW.  If all of that fails, fail the
load.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>